pizza_project/requirements.txt
Django==2.2.5
django-filter==2.3.0
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2020.4
sqlparse==0.2.4